refactor: 메인 컨트롤러 삭제
