+++ src/main/java/kr/codesqaud/cafe/config/MvcConfig.java
@@ -2,6 +2,7 @@ package kr.codesqaud.cafe.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
+import org.springframework.web.servlet.config.annotation.ViewControllerRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 @Configuration
@@ -14,10 +15,15 @@ public class MvcConfig implements WebMvcConfigurer {
     }
 
     @Override
+    public void addViewControllers(ViewControllerRegistry registry) {
+        registry.addRedirectViewController("/","/posts");
+    }
+
+    @Override
     public void addInterceptors(InterceptorRegistry registry) {
         registry.addInterceptor(signInInterceptor)
             .addPathPatterns("/**")
-            .excludePathPatterns("/", "/members/sign-up", "/sign-in", "/sign-out",
+            .excludePathPatterns("/", "/posts", "/members/sign-up", "/sign-in", "/sign-out",
                 "/css/**", "/error/**", "/js/**");
     }
 }
