--- src/test/java/kr/codesqaud/cafe/controller/MainControllerTest.java
@@ -1,46 +1,0 @@
-package kr.codesqaud.cafe.controller;
-
-import static org.mockito.BDDMockito.given;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
-import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.model;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.view;
-
-import java.util.Collections;
-import kr.codesqaud.cafe.service.PostService;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.http.MediaType;
-import org.springframework.test.web.servlet.MockMvc;
-
-@WebMvcTest(MainController.class)
-public class MainControllerTest {
-
-    @Autowired
-    private MockMvc mockMvc;
-
-    @MockBean
-    private PostService postService;
-
-    @DisplayName("ROOT URL을 접근하면 게시글 목록 페이지로 이동한다")
-    @Test
-    void showMain() throws Exception {
-        // given
-        given(postService.findAll()).willReturn(Collections.emptyList());
-
-        // when
-
-        // then
-        mockMvc.perform(get("/"))
-            .andExpect(status().isOk())
-            .andExpect(content().contentTypeCompatibleWith(MediaType.TEXT_HTML))
-            .andExpect(view().name("index"))
-            .andExpect(model().attributeExists("postResponses"))
-            .andDo(print());
-    }
-}
